requirements.txt
requests==2.31.0
pyjwt==2.8.0
protobuf==5.27.2
psutil==5.9.5
urllib3==2.0.7
pycryptodome==3.20.0
telebot==0.0.5
protobuf-decoder==1.0